地面の画像を作成。背景スクロールで隙間が生じる不具合を修正。

+++ toritoma/PlayingScene/AKBack.h
@@ -48,6 +48,6 @@ struct AKBackDef {
 }
 
 // 背景生成処理
-- (void)createBackType:(NSInteger)type x:(NSInteger)x y:(NSInteger)y parent:(CCNode *)parent;
+- (void)createBackType:(NSInteger)type x:(float)x y:(float)y parent:(CCNode *)parent;
 
 @end
